create-a-payment-link/requirements.txt
stripe==7.0.0
python-dotenv==0.21.1
